[BinGraph] Remove edges after exit syscall
This heuristic evaluates the basic block which contains syscall
instruction, and investigates the value of syscall number indicator. So
if the value maches with the exit syscall, we think that it is exit
syscall. If the syscall number is not set within the basic block, we
decide that we cannot determine whether it is exit or not.