[BinFile] Change the addr validity check routines
Some segments in memory (when loaded) have uninitialized regions that do
not necessarily exist in the binary file. Thus, VM size can be bigger
than the segment stored in the file. Previously, we only considered
regions stored in a *section* as valid because sections (and their whole
bodies) are typically stored in a file. However, this can limit our
ability as there can be some regions in a segment that will be loaded
into a memory region even though they are not part of a section.
Indeed, Mach-O files often have some file contents that are within a
segment, but not within a section, and we were not able to see such
regions under B2R2 as their addresses were considered *invalid*.

This commit fixes the above issue by always considering segments when
deciding the validity of addresses. Several FileInfo APIs are addeded
including `IsInFileAddr`, `IsInFileRange`, `IsValidRange`, and
`GetNotInFileIntervals`.