[FrontEnd] Remove ParsingContext
Remove ParsingContext and make Parser (Parsing process) mutable from now
on. When parsing in parallel, we should create a new Parser for each
thread. This is to give more flexibility for ISA-specific parsing
routines while producing less garbages. Note this is one of the few
places where we give up the immutability.