[MiddleEnd] Fix rollback sync problem
When a function is reloaded due to the force-finish mechanism, it may
find a previously unseed jump table, and the jump table can then have a
conflict with a previously known table. In this case, we will perform
a rollback by modifying the potential endpoint of the prev table.
However, we should also carefully modify the confirmed endpoint of the
prev table in order to avoid infinite recursion.