[MiddleEnd] Fix reload logic
The fix (cae032b1fea4301bb7939ebe588928f7c08caed4) was insufficient
because it only considers force-finished builders. It is safer to
restart the whole building process when the builder contains at least
one jump table because jump table identification itself can be wrong
when a CFG is under-approximated because a bogus dataflow can mess up
the table identification process, too.

So, we restart the whole process whenever a builder contains a jump
table to be safe.